+++ flink-table/flink-table-api-java/src/main/java/org/apache/flink/table/expressions/resolver/rules/ResolveCallByArgumentsRule.java
@@ -38,7 +38,6 @@ import org.apache.flink.table.functions.BuiltInFunctionDefinitions;
 import org.apache.flink.table.functions.FunctionDefinition;
 import org.apache.flink.table.functions.FunctionIdentifier;
 import org.apache.flink.table.functions.FunctionKind;
-import org.apache.flink.table.functions.ModelSemantics;
 import org.apache.flink.table.functions.ScalarFunctionDefinition;
 import org.apache.flink.table.functions.TableAggregateFunctionDefinition;
 import org.apache.flink.table.functions.TableFunctionDefinition;
@@ -650,12 +649,6 @@ final class ResolveCallByArgumentsRule implements ResolverRule {
                             DataTypeUtils.removeTimeAttribute(tableRef.getOutputDataType()),
                             staticArg);
             return Optional.of(semantics);
-        }
-
-        @Override
-        public Optional<ModelSemantics> getModelSemantics(int pos) {
-            // TODO: Add ModelReferenceExpression checks and TableApiModelSemantics
-            return Optional.empty();
         }
 
         @Override
