+++ flink-table/flink-table-planner/src/main/java/org/apache/flink/table/planner/plan/nodes/physical/stream/StreamPhysicalProcessTableFunction.java
@@ -248,12 +248,11 @@ public class StreamPhysicalProcessTableFunction extends AbstractRelNode
         final BridgingSqlFunction.WithTableFunction function =
                 (BridgingSqlFunction.WithTableFunction) rexCall.getOperator();
 
-        if (!function.getTypeInference().allowSystemArguments()) {
+        if (function.getTypeInference().disableSystemArguments()) {
             // Disabling uid and time attributes for process table functions with implementation
             // is not supported for now. It can only be disabled for syntax purpose: for example
             // it's disabled for ML_PREDICT which is not processed by this rule.
-            throw new ValidationException(
-                    "Disabling uid/time attributes is not supported for PTF.");
+            throw new ValidationException("Disabling system arguments is not supported for PTF.");
         }
     }
